requirements.txt
backcall==0.2.0
click==8.0.4
colorama==0.4.4
decorator==5.1.1
dlib==19.8.1
entrypoints==0.4
face-recognition==1.3.0
face-recognition-models==0.3.0
face-recognize==0.2.0
importlib-metadata==4.8.3
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
jedi==0.17.2
jupyter-client==7.1.2
jupyter-core==4.9.2
nest-asyncio==1.5.5
numpy==1.19.5
opencv-python==4.5.5.64
parso==0.7.1
pickleshare==0.7.5
Pillow==8.4.0
pip==21.3.1
prompt-toolkit==3.0.29
Pygments==2.11.2
python-dateutil==2.8.2
pywin32==303
pyzmq==22.3.0
setuptools==39.0.1
six==1.16.0
tornado==6.1
traitlets==4.3.3
typing_extensions==4.1.1
wcwidth==0.2.5
zipp==3.6.0